requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
Flask==3.1.1
flask-cors==6.0.1
future==1.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
netmiko==4.4.0
ntc_templates==7.9.0
packaging==25.0
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
pyserial==3.5
PyYAML==6.0.2
scp==0.15.0
six==1.17.0
textfsm==1.1.3
Werkzeug==3.1.3
